completed functionality and removed local imports of header
