+++ src/components/pages/AddChild/RenderAddChild.js
@@ -1,9 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-
 import { Layout, Typography } from 'antd';
-
-import ParentNavTopBar from '../../common/ParentNavTop';
 import ParentDashboardBack from '../../common/ParentDashboardBack';
 import ChildForm from '../../common/ChildForm';
 
@@ -12,7 +9,6 @@ const { Title } = Typography;
 const RenderAddChild = props => {
   return (
     <Layout className="add-child">
-      <ParentNavTopBar />
       <Layout className="content">
         <div className="top-section">
           <ParentDashboardBack />
