+++ src/components/pages/ParentContact/ContactUsContainer.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import RenderContactUs from './RenderContactUs';
-import ParentNavTopBar from '../../common/ParentNavTop';
 const ParentContactContainer = () => {
   return (
     <>
-      <ParentNavTopBar />
       <RenderContactUs />
     </>
   );
